requirements.txt
mlflow
opencv-python
Pillow
Django
tensorflow
django-filter
django-cors-headers
djangorestframework
openai
langchain
opencv-contrib-python
requests
gtts
pydub
Pillow